feat(proxy): refactor expert routing and use grouped dropdown
