+++ src/locales/en.json
@@ -495,6 +495,7 @@
             "route_to": "Route To",
             "haiku_optimization_tip": "Claude CLI uses {{model}} for background tasks by default. Mapping to a cheap Flash model can save ~95% cost.",
             "haiku_optimization_btn": "Quick Optimize",
+            "select_target_model": "Select target model",
             "money_saving_tip": "Savings Tip:"
         },
         "multi_protocol": {
